requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
cloudinary==1.42.1
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==5.1.5
django-allauth==65.3.1
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
gunicorn==20.1.0
psycopg2==2.9.10
whitenoise==6.8.2
